Player Attack & Sound
검 기본스킬 몬스터 관통하게 수정 완료

활 숙련도 스킬 구성 완료 - 나무 스킬중 나무별 색상에 맞는 파티클 추가함

도끼 숙련도 스킬 차징 기능 추가 완료

검, 도끼, 활 기본공격 및 기본스킬 사운드 추가 완료

플레이어 점프 및 슬라이딩 사운드 추가 완료

+++ Assets/pak0823/red cap project/Script/Arrow.cs
@@ -8,7 +8,7 @@ public class Arrow : MonoBehaviour
     public Enemy enemy;
     public LayerMask islayer; // 충돌 감지를 할 레이어
     public Transform pos; // 화살 위치 정보
-    public int Dmg = 5; //대미지 변수, 몬스터가 피격시 화살 데미지값을 받기 위해
+    public float Dmg = 5; //대미지 변수, 몬스터가 피격시 화살 데미지값을 받기 위해
     public float speed = 15f; // 화살 이동 속도
     private bool isSkill = false; // 스킬 사용 여부
     private Vector3 moveDirection = Vector3.right; // 화살이 나가는 방향
